+++ CHANGELOG.md
@@ -1,5 +1,3 @@
 ## CHANGELOG
 
-1. Better post selftext formatting, no more clutter.
-2. back to using ref on login user input.
-3. Home sub selection input now doesnt go to not found page on empty submit.
+1. added partial markdown support for postbody.